requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
email-validator==2.2.0
openpyxl==3.1.5
reportlab==4.2.5
python-dotenv==1.0.1
WTForms==3.2.1

